feat: Custom ScrollView and FlatList with styled scrollable indicator
